+++ lib/vania.dart
@@ -20,6 +20,7 @@ export 'src/websocket/websocket_event.dart';
 
 export 'src/route/router.dart';
 export 'src/route/route.dart';
+export 'src/route/middleware/throttle.dart';
 
 export 'src/cache/cache_driver.dart';
 export 'src/cache/cache.dart';
@@ -37,6 +38,8 @@ export 'src/database/mysql_driver.dart';
 export 'src/database/postgresql_driver.dart';
 export 'src/database/migration.dart';
 export 'src/enum/column_index.dart';
+export 'src/extensions/pagination_extension.dart';
+export 'src/extensions/simple_paginate_extension.dart';
 export 'package:eloquent/src/query/query_builder.dart';
 
 export 'src/authentication/authentication.dart';